General code refactor and added more unit tests

+++ HotelBookingSystem.Application/Services/GuestReviewService.cs
@@ -4,7 +4,6 @@ using HotelBookingSystem.Domain.Entities;
 using HotelBookingSystem.Domain.Interfaces.Repository;
 using System.Security.Claims;
 using Microsoft.AspNetCore.Http;
-using HotelBookingSystem.Infrastructure.Repository;
 
 
 namespace HotelBookingSystem.Application.Services
@@ -26,10 +25,10 @@ namespace HotelBookingSystem.Application.Services
             _userRepository = userRepository;
         }
 
-        public async Task<GuestReviewResponse> AddReviewAsync(GuestReviewRequest reviewRequest)
+        public async Task<GuestReviewResponse> AddReviewAsync(int hotelId, GuestReviewRequest reviewRequest)
         {
 
-            var hotel = await _hotelRepository.GetByIdAsync(reviewRequest.HotelId);
+            var hotel = await _hotelRepository.GetByIdAsync(hotelId);
 
             if (hotel == null)
                 throw new KeyNotFoundException("Hotel not found");
@@ -37,35 +36,39 @@ namespace HotelBookingSystem.Application.Services
             var userId = int.Parse(_httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? throw new UnauthorizedAccessException("User not found"));
 
             var review = _mapper.Map<GuestReview>(reviewRequest);
+            review.HotelId = hotelId;
 
             review.UserId = userId;
             review.User = await _userRepository.GetByIdAsync(userId);
 
             await _guestReviewRepository.AddAsync(review);
 
-            await UpdateHotelRating(reviewRequest.HotelId, hotel);
+            await UpdateHotelRating(hotel);
 
             return _mapper.Map<GuestReviewResponse>(review);
         }
 
-        public async Task DeleteReviewAsync (int reviewId)
+        public async Task DeleteReviewAsync (int hotelId, int reviewId)
         {
             var review = await _guestReviewRepository.GetByIdAsync(reviewId);
             if (review == null)
                 throw new KeyNotFoundException("Review not found");
+
+            if (review.HotelId != hotelId)
+                throw new ArgumentException("Invalid hotel ID");
 
             var hotel = await _hotelRepository.GetByIdAsync(review.HotelId);
             if (hotel == null)
                 throw new KeyNotFoundException("Hotel not found");
             await _guestReviewRepository.DeleteAsync(reviewId);
 
-            await UpdateHotelRating(review.HotelId, hotel);
+            await UpdateHotelRating(hotel);
 
         }
 
-        private async Task UpdateHotelRating(int hotelId, Hotel hotel)
+        private async Task UpdateHotelRating(Hotel hotel)
         {
-            var reviews = await _guestReviewRepository.GetReviewsByHotelIdAsync(hotelId);
+            var reviews = await _guestReviewRepository.GetReviewsByHotelIdAsync(hotel.HotelId);
             var averageRating = reviews.Any() ? reviews.Average(r => r.Rating) : 0;
 
             if (hotel != null)
